Add post body test

+++ src/worker/worker.c
@@ -147,7 +147,7 @@ int clientSendPacketList(struct client *c)
         removeListNode(c->res_buf, node);
     }
     freeListIterator(iter);
-    return allsend;
+    return (int)allsend;
 }
